Use GEM_HOST_API_KEY
`gem help push` reads

> [...]
>
> The push command will use ~/.gem/credentials to authenticate to a server,
> but you can use the RubyGems environment variable GEM_HOST_API_KEY to set
> the api key to authenticate.

So using the ENV variable that is already available, we don't need the hassle
of creating a temporary file.

Another thing is, in Rubygems source code I found that we could use
GEM_HOST_OTP_CODE instead of passing the `--otp` flag. But since the otp value
is currently not shared outside the "Publish to RubyGems" step, storing that
value in an ENV variable might actually decrease security.